+++ scripts/migrate_otrs2kix.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 # --
-# Copyright (C) 2006-2016 c.a.p.e. IT GmbH, http://www.cape-it.de
+# Copyright (C) 2006-2017 c.a.p.e. IT GmbH, http://www.cape-it.de
 #
 # written/edited by:
 # * Rene(dot)Boehm(at)cape(dash)it(dot)de
@@ -110,14 +110,7 @@ echo stopping cronjobs
 /opt/kix/bin/Cron.sh stop $APACHEUSER 2>&1 >> $LOGFILE
 
 echo stopping daemon
-sudo -u $APACHEUSER /opt/kix/bin/otrs.Daemon.pl stop --force 1>> $LOGFILE 2>&1
-
-# extract all packages for later use
-set -e # exit on every error
-echo extracting packages
-mkdir -p $TMP_PATH/kix_migration
-chmod -R g+w,o+w $TMP_PATH/kix_migration
-sudo -u $APACHEUSER /opt/kix/bin/otrs.Console.pl Admin::KIX::Migration::ExtractPackages --dir $TMP_PATH/kix_migration 2>> $LOGFILE
+sudo -u $APACHEUSER /opt/kix/bin/kix.Daemon.pl stop --force 1>> $LOGFILE 2>&1
 
 case $KIX_DBMS in
     mysql)
@@ -171,6 +164,7 @@ case $KIX_DBMS in
         echo "  importing database dump"
         export PGPASSWORD=$KIX_DBPW;export PGOPTIONS='--client-min-messages=warning';psql -q -h localhost -U kix $KIX_DB -f $TMP_PATH/kix_migration/kix_db.dmp 2>> $LOGFILE >> $LOGFILE
         ;;
+
 esac
 
 # migrate config
@@ -181,35 +175,27 @@ for FILE in ZZZAuto.pm ZZZACL.pm ZZZProcessManagement.pm; do
     fi
 done
 
-# reinstall opm files
-echo "re-installing packages (depending on the size of your OTRS database this might take a long time)"
-for FILE in `ls $TMP_PATH/kix_migration/*.opm | sort -r`; do
-    echo -n "."
-    sudo -u $APACHEUSER bash -c "/opt/kix/bin/otrs.Console.pl Admin::Package::Install --force $FILE >> $LOGFILE 2>&1"
-done
-echo
-
 # disable hard error handling
 set +e
 
-# reinstall KIX4OTRS to activate relevant configurations
-sudo -u $APACHEUSER bash -c "/opt/kix/bin/otrs.Console.pl Admin::Package::Reinstall --force KIX4OTRS >> $LOGFILE 2>&1"
+# upgrade database and cleanup obsolete packages
+sudo -u $APACHEUSER bash -c "/opt/kix/scripts/database/update/kix-upgrade-to-17.pl >> $LOGFILE 2>&1"
 
 # clear user skins
-sudo -u $APACHEUSER bash -c "/opt/kix/bin/otrs.Console.pl Admin::User::ClearPreferences --key UserSkin >> $LOGFILE 2>&1"
+sudo -u $APACHEUSER bash -c "/opt/kix/bin/kix.Console.pl Admin::User::ClearPreferences --key UserSkin >> $LOGFILE 2>&1"
 
 # remove double entries in package_repository due to force install
-sudo -u $APACHEUSER bash -c "/opt/kix/bin/otrs.Console.pl Admin::KIX::CleanupPackageRepository >> $LOGFILE 2>&1"
+sudo -u $APACHEUSER bash -c "/opt/kix/bin/kix.Console.pl Admin::KIX::CleanupPackageRepository >> $LOGFILE 2>&1"
 
 # remove temporary directory
 rm -rf $TMP_PATH/kix_migration
 
 echo rebuilding config
-sudo -u $APACHEUSER bash -c "/opt/kix/bin/otrs.Console.pl Maint::Config::Rebuild 2>&1 >> $LOGFILE"
+sudo -u $APACHEUSER bash -c "/opt/kix/bin/kix.Console.pl Maint::Config::Rebuild 2>&1 >> $LOGFILE"
 
 echo deleting caches
-sudo -u $APACHEUSER bash -c "/opt/kix/bin/otrs.Console.pl Maint::Cache::Delete 2>&1 >> $LOGFILE"
-sudo -u $APACHEUSER bash -c "/opt/kix/bin/otrs.Console.pl Maint::Loader::CacheCleanup 2>&1 >> $LOGFILE"
+sudo -u $APACHEUSER bash -c "/opt/kix/bin/kix.Console.pl Maint::Cache::Delete 2>&1 >> $LOGFILE"
+sudo -u $APACHEUSER bash -c "/opt/kix/bin/kix.Console.pl Maint::Loader::CacheCleanup 2>&1 >> $LOGFILE"
 
 echo
 echo "*****************************************************************************************"